Refactor code by reusing functions in Utils and Helper extension and class
