+++ RouteMe/Utils.swift
@@ -22,7 +22,27 @@ extension UIView {
         
         self.addSubview(imageViewBackground)
         self.sendSubviewToBack(imageViewBackground)
-    }}
+    }
+    
+    func startASpinner() -> UIView {
+        var messageFrame = UIView()
+        var activityIndicator = UIActivityIndicatorView()
+        messageFrame = UIView(frame: CGRect(x: self.frame.midX - 25, y: self.frame.midY - 25 , width: 50, height: 50))
+        messageFrame.layer.cornerRadius = 15
+        messageFrame.backgroundColor = UIColor(white: 0, alpha: 0.7)
+        activityIndicator = UIActivityIndicatorView(activityIndicatorStyle: UIActivityIndicatorViewStyle.White)
+        activityIndicator.frame = CGRect(x: 0, y: 0, width: 50, height: 50)
+        activityIndicator.startAnimating()
+        messageFrame.addSubview(activityIndicator)
+        self.addSubview(messageFrame)
+        return messageFrame
+    }
+    
+    func stopSpinner(spinnerFrame: UIView) {
+        spinnerFrame.removeFromSuperview()
+    }
+    
+}
 
 extension UIViewController {
     func hideKeyboardWhenTappedAround() {
@@ -33,4 +53,19 @@ extension UIViewController {
     func dismissKeyboard() {
         view.endEditing(true)
     }
+    
+    func alert(title: String, message: String, buttonText: String) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertControllerStyle.Alert);
+        let alertButton = UIAlertAction(title: buttonText, style: UIAlertActionStyle.Default, handler: {(actionTarget: UIAlertAction) in })
+        alert.addAction(alertButton)
+        showViewController(alert, sender: self);
+    }
+    
+    func redirectToMainView() {
+        dispatch_async(dispatch_get_main_queue(), { () -> Void in
+            let viewController:UIViewController = UIStoryboard(name: "Main", bundle: nil).instantiateViewControllerWithIdentifier("Home")
+            self.presentViewController(viewController, animated: true, completion: nil)
+        })
+    }
 }
+
