+++ RouteMe/helpers.swift
@@ -1,0 +1,58 @@
+//
+//  Helpers.swift
+//  RouteMe
+//
+//  Created by Hesham Massoud on 21/05/16.
+//  Copyright © 2016 Hesham Massoud. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class Helper{
+    static func validateEmail(email: String) -> Bool {
+        let emailRegEx = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}"
+        let emailEvaluator = NSPredicate(format:"SELF MATCHES %@", emailRegEx)
+        return emailEvaluator.evaluateWithObject(email)
+    }
+    
+    static func validatePassword(password: String) -> Bool {
+        return password.characters.count > 1
+    }
+    
+    static func passwordsMatch(password: String, confirmPassword: String) -> Bool {
+        return password == confirmPassword
+    }
+    
+    static func alertRequestError(responseJSON: NSDictionary, viewController: UIViewController) {
+        let FIRST_FIELD_ERROR_INDEX = 0
+        let FIRST_FIELD_ERROR_ENTRY = responseJSON["fieldErrors"]![FIRST_FIELD_ERROR_INDEX] as AnyObject
+        let errorMessage = FIRST_FIELD_ERROR_ENTRY["message"] as! String
+        let errorField = FIRST_FIELD_ERROR_ENTRY["field"] as! String
+        viewController.alert(errorField, message: errorMessage, buttonText: "OK")
+    }
+    
+    static func loginUser(user: User, viewController: UIViewController) {
+        rememberUser(user)
+        redirectToViewController(viewController, targetViewControllerId: "Home")
+    }
+    
+    static func rememberUser(user: User) {
+        let hasLoginKey = NSUserDefaults.standardUserDefaults().boolForKey("isLoggedIn")
+        if hasLoginKey == false {
+            NSUserDefaults.standardUserDefaults().setBool(true, forKey: "isLoggedIn")
+            NSUserDefaults.standardUserDefaults().setValue(user.id, forKey: "loggedInId")
+            NSUserDefaults.standardUserDefaults().setValue(user.username, forKey: "loggedInUsername")
+            NSUserDefaults.standardUserDefaults().setValue(user.email, forKey: "loggedInEmail")
+        }
+    }
+    
+    static func redirectToViewController(viewController: UIViewController, targetViewControllerId: String) {
+        dispatch_async(dispatch_get_main_queue(), { () -> Void in
+            let targetViewController:UIViewController = UIStoryboard(name: "Main", bundle: nil).instantiateViewControllerWithIdentifier(targetViewControllerId)
+            viewController.presentViewController(targetViewController, animated: true, completion: nil)
+        })
+    }
+    
+    
+}
